api rest cambiada

+++ src/main/java/com/example/GeniusApp/Controllers/CommentRESTController.java
@@ -43,7 +43,6 @@ public class CommentRESTController {
     @PostMapping("/songs/{id}")
     public ResponseEntity<Comment> createComment(HttpServletRequest request, @PathVariable long id, @RequestBody Comment comment){
         Song song = songService.getSong(id);
-
         commentService.addComment(song,Sanitizers.FORMATTING.sanitize(comment.getText()), request.getUserPrincipal().getName());
         return new ResponseEntity<>(comment, HttpStatus.CREATED);
     }
@@ -62,12 +61,17 @@ public class CommentRESTController {
 
     @Transactional
     @DeleteMapping("/songs/{sid}/{cid}")
-    public void deleteComment(@PathVariable long sid, @PathVariable long cid) {
+    public int deleteComment(HttpServletRequest request, @PathVariable long sid, @PathVariable long cid) {
         Song song = songService.getSong(sid);
         Comment comment = commentService.getComment(cid);
-        song.getComments().size();
-        song.getComments().remove(comment);
-        commentService.removeComment(cid,sid);
+        if (request.getUserPrincipal().getName().equals(comment.getOwner()) || request.isUserInRole("ADMIN")){
+            song.getComments().size();
+            song.getComments().remove(comment);
+            commentService.removeComment(cid,sid);
+            return 1;
+        }else{
+            return 0;
+        }
     }
 
     @Transactional
